Use MyList/MyArrayList instead of ArrayList
Add MyList and MyArrayList, an interface and class (respectively) that
extend List and ArrayList (respectively). They add overloads of get,
set, and remove that assume the index is size() - 1. Use MyList and
MyArrayList in place of List and ArrayList in GrowingTreeMaze. Because
these classes eliminate the use for the pop and last methods in
ArrayUtil, remove these methods.