Implement SolutionMaze in GrowingTreeMaze
Add an instance variable, solution, to GrowingTreeMaze in order to hold
the solution once it is computed. Extend
ReversibleGeneratingSolutionMaze in GrowingTreeMaze and implement its
showSolution and hideSolution methods. Change the getFrontChar,
getLeftChar, and getCenterChar methods in ArrayMaze from private to
protected, and override them in GrowingTreeMaze so that the toString
representation of a GrowingTreeMaze shows the solution after
showSolution is called.